pyproject.toml
[tool.poetry]
name = "todoist-digest"
version = "0.1.0"
description = "Generate a digest of comments that occured since the last checked in date. Useful for including in an email."
authors = ["Michael Bianco <mike@mikebian.co>"]
license = "MIT"
readme = "README.md"
keywords = ["todoist", "digest", "task", "management"]

[tool.poetry.dependencies]
python = "^3.12"
todoist-api-python = "^2.1.3"
click = "^8.1.7"
pyjson5 = "^1.6.4"
funcy-pipe = "^0.4.0"
# funcy-pipe = { develop = true, path = "../../python/funcy-pipe" }
python-dateutil = "^2.8.2"
backoff = "^2.2.1"

[tool.poetry.group.dev.dependencies]
isort = "5.11.5"
black = "^23.10.1"
pylint = "^3.0.2"

# --float-to-top
[tool.isort]
profile = "black"
float_to_top = true

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
